# [POC] Improve TypeScript Theme Components Performance