--- a/packages/mui-lab/src/themeAugmentation/components.ts
+++ b/packages/mui-lab/src/themeAugmentation/components.ts
@@ -2,8 +2,8 @@ import {
   ComponentsProps,
   ComponentsOverrides,
   ComponentsVariants,
-  Theme as MuiTheme,
-} from '@mui/material/styles';
+} from '@mui/material/deprecatedTypes';
+import { Theme as MuiTheme } from '@mui/material/styles';
 
 type Theme = Omit<MuiTheme, 'components'>;
 
@@ -68,6 +68,6 @@ export interface LabComponents {
   };
 }
 
-declare module '@mui/material/styles' {
+declare module '@mui/material/deprecatedTypes' {
   interface Components extends LabComponents {}
 }