--- a/packages/mui-material/perf-test/test-createTheme.ts
+++ b/packages/mui-material/perf-test/test-createTheme.ts
@@ -1,0 +1,16 @@
+import { ButtonTheme } from '@mui/material/Button';
+import { createTheme } from '../src/styles';
+
+declare module '@mui/material/styles' {
+  interface ThemeComponents extends ButtonTheme {}
+}
+
+export default createTheme({
+  components: {
+    MuiButton: {
+      defaultProps: {
+        disableElevation: true,
+      },
+    },
+  },
+});