--- a/packages/mui-material/perf-test/tsconfig.json
+++ b/packages/mui-material/perf-test/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../tsconfig.json",
+  "include": ["test-createTheme.ts"],
+  "compilerOptions": {
+    "moduleResolution": "Bundler"
+  }
+}