--- a/packages/mui-material/src/styles/index.js
+++ b/packages/mui-material/src/styles/index.js
@@ -1,5 +1,4 @@
 export { default as THEME_ID } from './identifier';
-export { default as adaptV4Theme } from './adaptV4Theme';
 export {
   hexToRgb,
   rgbToHex,