--- a/packages/mui-material/src/stylesDeprecated/index.ts
+++ b/packages/mui-material/src/stylesDeprecated/index.ts
@@ -1,0 +1,6 @@
+export { default as adaptV4Theme } from './adaptV4Theme';
+export { DeprecatedThemeOptions } from './adaptV4Theme';
+export { ComponentsProps, ComponentsPropsList } from './props';
+export { ComponentsVariants } from './variants';
+export { ComponentsOverrides, ComponentNameToClassKey } from './overrides';
+export { Components } from './components';